random stuff

+++ Source/Engine/Assets/AssetsSystem/AssetsSystem.cpp
@@ -11,11 +11,26 @@ Bool AssetsSystem::Initialize(const AssetsSystemInitializer& initializer)
 {
 	m_contentPath = initializer.contentPath;
 
+	m_ddc.Initialize({ .path = initializer.ddcPath });
+
 	return true;
 }
 
 void AssetsSystem::Shutdown()
 {
+	lib::LockGuard lock(m_assetsSystemLock);
+
+	if (!m_loadedAssets.empty())
+	{
+		const lib::DynamicArray<AssetHandle> loadedAssets = GetLoadedAssetsList_Locked();
+
+		for (const AssetHandle& asset : loadedAssets)
+		{
+			SPT_LOG_ERROR(AssetsSystem, "Asset was not unloaded! Name: {}", asset->GetName().ToString());
+		}
+
+		SPT_CHECK_NO_ENTRY_MSG("Not all assets were unloaded before shutdown!");
+	}
 }
 
 CreateResult AssetsSystem::CreateAsset(const AssetInitializer& initializer)
@@ -33,9 +48,20 @@ CreateResult AssetsSystem::CreateAsset(const AssetInitializer& initializer)
 
 	const AssetHandle assetInstance = CreateAssetInstance(initializer);
 
+	if (!assetInstance.IsValid())
+	{
+		return CreateResult(ECreateError::FailedToCreateInstance);
+	}
+
+	assetInstance->PostCreate();
+
+	SPT_CHECK_MSG(assetInstance->GetBlackboard().GetUnloadedTypes().empty(), "Asset data was unloaded during creation. This is not allowed.");
+
 	SaveAssetImpl(assetInstance);
 
 	m_loadedAssets[initializer.path] = assetInstance.Get();
+
+	assetInstance->PostInitialize();
 
 	return CreateResult(std::move(assetInstance));
 }
@@ -64,10 +90,10 @@ LoadResult AssetsSystem::LoadAsset(const ResourcePath& path)
 	AssetInstanceData assetData = std::move(ReadAssetData(fullPath));
 
 	const AssetHandle assetInstance = CreateAssetInstance(AssetInitializer
-													{
-														.type = assetData.type,
-														.path = path,
-													});
+														  {
+															  .type = assetData.type,
+															  .path = path,
+														  });
 
 	if (!assetInstance.IsValid())
 	{
@@ -77,6 +103,10 @@ LoadResult AssetsSystem::LoadAsset(const ResourcePath& path)
 	assetInstance->AssignData(std::move(assetData));
 
 	m_loadedAssets[path] = assetInstance.Get();
+
+	assetInstance->PostLoad();
+
+	assetInstance->PostInitialize();
 
 	return LoadResult(AssetHandle(std::move(assetInstance)));
 }
@@ -102,7 +132,23 @@ EDeleteResult AssetsSystem::DeleteAsset(const ResourcePath& path)
 		return EDeleteResult::DoesNotExist;
 	}
 
-	std::filesystem::remove(fullPath);
+	AssetInstanceData assetData;
+
+	{
+		lib::LockGuard lock(m_assetsSystemLock);
+
+		if (IsLoaded_Locked(path))
+		{
+			return EDeleteResult::StillReferenced;
+		}
+
+		assetData = ReadAssetData(fullPath);
+
+		std::filesystem::remove(fullPath);
+	}
+
+	AssetFactory& factory = AssetFactory::GetInstance();
+	factory.DeleteAsset(assetData.type, *this, path, assetData);
 
 	return EDeleteResult::Success;
 }
@@ -123,7 +169,7 @@ Bool AssetsSystem::IsLoaded(const ResourcePath& path) const
 {
 	lib::LockGuard lock(m_assetsSystemLock);
 
-	return m_loadedAssets.contains(path);
+	return IsLoaded_Locked(path);
 }
 
 lib::DynamicArray<AssetHandle> AssetsSystem::GetLoadedAssetsList() const
@@ -132,19 +178,13 @@ lib::DynamicArray<AssetHandle> AssetsSystem::GetLoadedAssetsList() const
 
 	lib::LockGuard lock(m_assetsSystemLock);
 
-	lib::DynamicArray<AssetHandle> result;
-	result.reserve(m_loadedAssets.size());
-
-	for (const auto& [path, asset] : m_loadedAssets)
-	{
-		result.emplace_back(asset);
-	}
-
-	return result;
+	return GetLoadedAssetsList_Locked();
 }
 
 void AssetsSystem::OnAssetUnloaded(AssetInstance& instance)
 {
+	instance.PreUnload();
+
 	lib::LockGuard lock(m_assetsSystemLock);
 
 	SPT_CHECK(instance.GetRefCount() == 0u);
@@ -157,6 +197,8 @@ void AssetsSystem::SaveAssetImpl(AssetHandle asset)
 	SPT_PROFILER_FUNCTION();
 
 	SPT_CHECK(asset.IsValid());
+
+	asset->PreSave();
 
 	AssetBlackboard& assetRuntimeData = asset->GetBlackboard();
 
@@ -193,6 +235,8 @@ void AssetsSystem::SaveAssetImpl(AssetHandle asset)
 	stream << data;
 
 	stream.close();
+
+	asset->PostSave();
 }
 
 AssetInstanceData AssetsSystem::ReadAssetData(const lib::Path& fullPath) const
@@ -217,7 +261,34 @@ AssetHandle AssetsSystem::CreateAssetInstance(const AssetInitializer& initialize
 {
 	AssetFactory& factory = AssetFactory::GetInstance();
 
-	return factory.CreateAsset(*this, initializer);
+	AssetHandle newAsset = factory.CreateAsset(*this, initializer);
+
+	if (newAsset.IsValid() && initializer.dataInitializer)
+	{
+		initializer.dataInitializer->InitializeNewAsset(*newAsset);
+	}
+
+	return newAsset;
+}
+
+lib::DynamicArray<AssetHandle> AssetsSystem::GetLoadedAssetsList_Locked() const
+{
+	SPT_PROFILER_FUNCTION();
+
+	lib::DynamicArray<AssetHandle> result;
+	result.reserve(m_loadedAssets.size());
+
+	for (const auto& [path, asset] : m_loadedAssets)
+	{
+		result.emplace_back(asset);
+	}
+
+	return result;
+}
+
+Bool AssetsSystem::IsLoaded_Locked(const ResourcePath& path) const
+{
+	return m_loadedAssets.contains(path);
 }
 
 } // spt::as
